Fixed a problem locally that was causing my DB connection to fail...
My ".env" file (which stores DB credentials and such) was not in the "server" directory where it needed to be.  It needs to be in the same directory as the app.js file which pulls in the dotenv module and runs its config() function.  At that time that module will look in the same directory for the ".env" file.  Since it was not there it failed to pull my DB credential values in.

Although the ".env" file is not in source control, I made better notes in the app.js file to help aleviate this confusion.